requirements.txt
pymupdf
pdfplumber
pytesseract
Pillow
transformers
torch
langchain
langchain-huggingface
langchain-community
faiss-cpu
